refactor: unified config loader, config-driven tier routing, central defaults
- Unified config loader: consolidated 12+ load paths into
  load_config_with_options / load_config_validated with LoadOptions
- Fixed ValidatedConfig.raw regression: split parse_from_resolved_path
  from load_from_resolved_path so raw captures pre-mutation state
- Config-driven tier routing: CascadeRouter now uses tier_map from
  ModelProfile.tier config instead of slug substring heuristics,
  threaded through 28+ call sites in helpers.rs and cascade_router.rs
- Central defaults: RetryPolicy::default() wired to use roko_core::defaults
  constants, renamed DEFAULT_RETRY_INITIAL_BACKOFF_MS to
  DEFAULT_RETRY_BASE_DELAY_MS with corrected values
- ModelTier added to ModelProfile, populated across 80+ models in roko.toml
- Fixed stale model slugs (haiku-3-5→haiku-4-5, opus-4→opus-4-6, etc.)
- Added atomic I/O module (roko_core::io) with sync/async variants
- Added config loader integration tests (discovery, roundtrip, validation)
- Migrated 10+ callsites from ad-hoc config loading to unified loader

+++ crates/roko-agent/src/dispatcher/alert.rs
@@ -11,7 +11,7 @@ use std::collections::HashMap;
 const DEFAULT_FAILURE_THRESHOLD: f64 = roko_core::defaults::DEFAULT_FAILURE_THRESHOLD;
 
 /// Minimum number of calls before alerts can fire.
-const DEFAULT_MIN_CALLS: u64 = roko_core::defaults::DEFAULT_ALERT_MIN_CALLS as u64;
+const DEFAULT_MIN_CALLS: u64 = roko_core::defaults::DEFAULT_ALERT_MIN_CALLS;
 
 // ─── AlertSeverity ───────────────────────────────────────────────────────
