+++ crates/roko-gate/src/property_test_gate.rs
@@ -32,16 +32,12 @@ pub const COUNTEREXAMPLE_DIGEST_LIMIT: usize = 2048;
 /// Default per-test case count (matches proptest's built-in default).
 ///
 /// Sourced from [`roko_core::defaults::DEFAULT_PROPTEST_CASES`].
-#[allow(clippy::cast_possible_truncation)] // Value is small (256), fits in u32.
-pub const DEFAULT_PROPTEST_CASES: u32 =
-    roko_core::defaults::DEFAULT_PROPTEST_CASES as u32;
+pub const DEFAULT_PROPTEST_CASES: u32 = roko_core::defaults::DEFAULT_PROPTEST_CASES;
 
 /// Default shrink iteration ceiling (matches proptest's built-in default).
 ///
 /// Sourced from [`roko_core::defaults::DEFAULT_MAX_SHRINK_ITERS`].
-#[allow(clippy::cast_possible_truncation)] // Value is small (4096), fits in u32.
-pub const DEFAULT_MAX_SHRINK_ITERS: u32 =
-    roko_core::defaults::DEFAULT_MAX_SHRINK_ITERS as u32;
+pub const DEFAULT_MAX_SHRINK_ITERS: u32 = roko_core::defaults::DEFAULT_MAX_SHRINK_ITERS;
 
 /// Rung 4 gate: run property/invariant tests and capture counterexamples.
 ///
